+++ infra/chunkedEncode/KBaseChunkedEncodeSessionManager.php
@@ -10,6 +10,8 @@
 	 */
 	abstract class KBaseChunkedEncodeSessionManager
 	{
+		const 	SessionStatsJSONLogPrefix = "SessionStatsJSON";
+		
 		protected $name = null;
 		protected $chunker = null;
 
@@ -208,14 +210,11 @@
 		 */
 		public function Merge()
 		{
-			$videoFilename = $this->chunker->getSessionName("video");
-			$oFh=fopen($videoFilename,"wb");
-			if($oFh===false){
-				return false;
-			}
 			$rv=$this->chunker->ConcatChunks();
-			fclose($oFh);
 			if($rv===false) {
+				KalturaLog::log($msgStr="FAILED to merge - missing concat'ed chunk video file, leaving!");
+				$this->returnMessages[] = $msgStr;
+				$this->returnStatus = KChunkedEncodeReturnStatus::MergeError;
 				return false;
 			}
 			
@@ -290,27 +289,19 @@
 				KalturaLog::log("source:".print_r($fileDtSrc,1));
 			}
 			
+			$sessionStats = new KChunkedEncodeSessionReportStats();
+			$sessionStats->num = $chunker->GetMaxChunks();
+			$sessionStats->lasted = $this->finishTime - $this->createTime;
 			{
 				KalturaLog::log("CSV,idx,startedAt,user,system,elapsed,cpu");
-				$userAcc = $systemAcc = $elapsedAcc = $cpuAcc = 0;
 				foreach($this->chunkExecutionDataArr as $idx=>$execData){
-					$userAcc+= $execData->user;
-					$systemAcc+= $execData->system;
-					$elapsedAcc+= $execData->elapsed;
-					$cpuAcc+= $execData->cpu;
+					$sessionStats->userCpu +=    $execData->user;
+					$sessionStats->systemCpu +=  $execData->system;
+					$sessionStats->elapsedCpu += $execData->elapsed;
 					
 					KalturaLog::log("CSV,$idx,$execData->startedAt,$execData->user,$execData->system,$execData->elapsed,$execData->cpu");
 				}
 				$cnt = $chunker->GetMaxChunks();
-				if($cnt>0) {
-					$userAvg 	= round($userAcc/$cnt,3);
-					$systemAvg 	= round($systemAcc/$cnt,3);
-					$elapsedAvg = round($elapsedAcc/$cnt,3);
-					$cpuAvg 	= round($cpuAcc/$cnt,3);
-				}
-				else
-					$userAvg = $systemAvg = $elapsedAvg = $cpuAvg = 0;
-
 			}
 			
 //			KalturaLog::log("LogFile: ".$chunker->getSessionName("log"));
@@ -335,7 +326,19 @@
 			KalturaLog::log("***********************************************************");
 			KalturaLog::log("* Session Summary (".date("Y-m-d H:i:s").")");
 			KalturaLog::log("* ");
-			KalturaLog::log("ExecutionStats:chunks($cnt),accum(elapsed:$elapsedAcc,user:$userAcc,system:$systemAcc),average(elapsed:$elapsedAvg,user:$userAvg,system:$systemAvg,cpu:$cpuAvg)");
+			if(isset($concurrencyLevel)) {
+				$val = round(end($this->concurrencyHistogram)/1000,2);
+				$idle = round($this->concurrencyHistogram[0]/1000,2);
+				$sessionStats->concurrency = $concurrencyLevel;
+				$sessionStats->concurrencyMax = key($this->concurrencyHistogram);
+				$sessionStats->concurrencyMaxTime = $val;
+				$sessionStats->concurrencyIdleTime = $idle;
+			}
+
+			KalturaLog::log("SessionStats:".$sessionStats->ToString());
+			$sessionStatsJson = json_encode($sessionStats);
+			KalturaLog::log("SessionStatsJSON $sessionStatsJson");
+			
 			if($sessionData->returnStatus==KChunkedEncodeReturnStatus::AnalyzeError){
 				$msgStr.= ",analyze:BAD";
 			}
@@ -476,20 +479,61 @@
 		/********************
 		 *
 		 */
-		protected static function getLogTail($logFilename, $size=5000)
+		public static function GetLogTail($logFilename, $size=5000)
 		{
 			$logTail = null;
-			if(file_exists($logFilename)) {
-				$fHd = fopen($logFilename,"r");
-				$fileSz = filesize($logFilename);
-				if($fileSz>$size)
-					fseek($fHd,$fileSz-$size);
-				$logTail = fread($fHd, $size);
-				fclose($fHd);
+			if(file_exists($logFilename)==false) 
+				return null;
+
+			$fHd=fopen($logFilename,"r");
+			if($fHd===false){
+				return null;
 			}
+			if(fseek($fHd , -$size, SEEK_END)!==0){
+				fclose($fHd);
+				return null;
+			}
+			if(($logTail=fread($fHd, $size))===false){
+				fclose($fHd);
+				return null;
+			}
+			fclose($fHd);
 			return $logTail;
 		}
-	
+
+		/********************
+		 * 
+		 */
+		public static function GetLogTailLines($logFilename, $size=5000)
+		{
+			$logTail = self::GetLogTail($logFilename, $size);
+			if($logTail==null)
+				return null;
+			$lines=preg_split("/\r\n|\n|\r/", $logTail);
+			return $lines;
+		}
+
+		/********************
+		 * 
+		 */
+		public static function GetSessionStatsJSON($logFilename, $size=10000)
+		{
+			KalturaLog::log("$logFilename, $size");
+			if($lines=self::GetLogTailLines($logFilename, $size)) {
+				foreach($lines as $idx=>$line) {
+					$prefix=self::SessionStatsJSONLogPrefix;
+					if(($pos=strpos($line,$prefix))===false)
+						continue;
+					KalturaLog::log("$line");
+					$jsonStr = substr($line, $pos+strlen($prefix));
+					$obj = json_decode($jsonStr);
+					KalturaLog::log("JSON:$jsonStr, object:".print_r($obj,1));
+					return $obj;
+				}
+			}
+			return null;
+		}
+
 		/********************
 		 *
 		 */