+++ plugins/reach/lib/kReachManager.php
@@ -30,7 +30,9 @@ class kReachManager implements kObjectChangedEventConsumer, kObjectCreatedEventC
 			"permission" => objectType::PERMISSION,
 			"permissionItem" => objectType::PERMISSIONITEM,
 			"userRole" => objectType::USERROLE,
-			"categoryEntry" => objectType::CATEGORY_ENTRY );
+			"categoryEntry" => objectType::CATEGORY_ENTRY,
+			"CaptionAsset" => CaptionAssetEventNotificationsPlugin::getEventNotificationEventObjectTypeCoreValue('CaptionAsset'),);
+
 		if (isset($mapObjectType[$eventObjectClassName]))
 		{
 			return $mapObjectType[$eventObjectClassName];
@@ -42,11 +44,11 @@ class kReachManager implements kObjectChangedEventConsumer, kObjectCreatedEventC
 	{
 		$existingCatalogItemIds = EntryVendorTaskPeer::retrieveExistingTasksCatalogItemIds($entryId, $allowedCatalogItemIds);
 		$catalogItemIdsToAdd = array_unique(array_diff($allowedCatalogItemIds, $existingCatalogItemIds));
-
+		$taskJobData = self::getTaskJobData($object);
 		foreach ($catalogItemIdsToAdd as $catalogItemIdToAdd)
 		{
 			//Pass the object Id as the context of the task
-			self::addEntryVendorTaskByObjectIds($entryId, $catalogItemIdToAdd, $profileId, $this->getContextByObjectType($object));
+			self::addEntryVendorTaskByObjectIds($entryId, $catalogItemIdToAdd, $profileId, $this->getContextByObjectType($object), $taskJobData);
 		}
 	}
 
@@ -235,6 +237,12 @@ class kReachManager implements kObjectChangedEventConsumer, kObjectCreatedEventC
 			return true;
 		}
 
+		if ($object instanceof CaptionAsset && in_array(assetPeer::STATUS, $modifiedColumns) && $object->getStatus() == asset::ASSET_STATUS_READY)
+		{
+			$event = new kObjectChangedEvent($object,$modifiedColumns);
+			return $this->shouldConsumeEvent($event);
+		}
+
 		return false;
 	}
 	
@@ -335,6 +343,12 @@ class kReachManager implements kObjectChangedEventConsumer, kObjectCreatedEventC
 				$this->consumeEvent($event);
 			}
 			return $this->checkAutomaticRules($object);
+		}
+
+		if ($object instanceof CaptionAsset && in_array(assetPeer::STATUS, $modifiedColumns) && $object->getStatus() == asset::ASSET_STATUS_READY)
+		{
+			$event = new kObjectChangedEvent($object,$modifiedColumns);
+			$this->consumeEvent($event);
 		}
 
 		return true;
@@ -453,7 +467,7 @@ class kReachManager implements kObjectChangedEventConsumer, kObjectCreatedEventC
 		return true;
 	}
 
-	public static function addEntryVendorTaskByObjectIds($entryId, $vendorCatalogItemId, $reachProfileId, $context = null)
+	public static function addEntryVendorTaskByObjectIds($entryId, $vendorCatalogItemId, $reachProfileId, $context = null, $taskJobData = null)
 	{
 		$entry = entryPeer::retrieveByPK($entryId);
 		$reachProfile = ReachProfilePeer::retrieveActiveByPk($reachProfileId);
@@ -502,6 +516,10 @@ class kReachManager implements kObjectChangedEventConsumer, kObjectCreatedEventC
 		$entryVendorTask = self::addEntryVendorTask($entry, $reachProfile, $vendorCatalogItem, false, $sourceFlavorVersion, $context, EntryVendorTaskCreationMode::AUTOMATIC);
 		if($entryVendorTask)
 		{
+			if ($taskJobData)
+			{
+				$entryVendorTask->setTaskJobData($taskJobData);
+			}
 			$entryVendorTask->save();
 		}
 		return $entryVendorTask;
@@ -662,6 +680,18 @@ class kReachManager implements kObjectChangedEventConsumer, kObjectCreatedEventC
 	{
 		if ($object instanceof categoryEntry)
 			return $object->getCategoryId();
+
+		return null;
+	}
+
+	protected static function getTaskJobData($object)
+	{
+		if($object instanceof CaptionAsset)
+		{
+			$taskJobData = new kTranslationVendorTaskData();
+			$taskJobData->captionAssetId = $object->getId();
+			return $taskJobData;
+		}
 
 		return null;
 	}